+++ packages/astro/test/dev-routing.test.js
@@ -151,9 +151,9 @@ describe('Development Routing', () => {
 			expect(response.status).to.equal(200);
 		});
 
-		it('200 when loading subpath root without trailing slash', async () => {
+		it('404 when loading subpath root without trailing slash', async () => {
 			const response = await fixture.fetch('/blog');
-			expect(response.status).to.equal(200);
+			expect(response.status).to.equal(404);
 		});
 
 		it('200 when loading another page with subpath used', async () => {