+++ packages/astro/test/test-utils.js
@@ -1,5 +1,5 @@
 import { execa } from 'execa';
-import fetch from 'node-fetch';
+import { polyfill } from '@astropub/webapi';
 import fs from 'fs';
 import { fileURLToPath } from 'url';
 import { loadConfig } from '../dist/core/config.js';
@@ -7,6 +7,12 @@ import dev from '../dist/core/dev/index.js';
 import build from '../dist/core/build/index.js';
 import preview from '../dist/core/preview/index.js';
 import os from 'os';
+
+// polyfill WebAPIs to globalThis for Node v12, Node v14, and Node v16
+polyfill(globalThis, {
+	exclude: 'window document',
+});
+
 /**
  * @typedef {import('node-fetch').Response} Response
  * @typedef {import('../src/core/dev/index').DevServer} DevServer
@@ -21,12 +27,13 @@ import os from 'os';
  * @property {(path: string) => Promise<string[]>} readdir
  * @property {() => Promise<DevServer>} startDevServer
  * @property {() => Promise<PreviewServer>} preview
+ * @property {() => Promise<void>} clean
  */
 
 /**
  * Load Astro fixture
  * @param {AstroConfig} inlineConfig Astro config partial (note: must specify projectRoot)
- * @returns {Fixture} The fixture. Has the following properties:
+ * @returns {Promise<Fixture>} The fixture. Has the following properties:
  *   .config     - Returns the final config. Will be automatically passed to the methods below:
  *
  *   Build
@@ -39,6 +46,9 @@ import os from 'os';
  *
  *   Preview
  *   .preview()        - Async. Starts a preview server. Note this can’t be running in same fixture as .dev() as they share ports. Also, you must call `server.close()` before test exit
+ *
+ *   Clean-up
+ *   .clean()          - Async. Removes the project’s dist folder.
  */
 export async function loadFixture(inlineConfig) {
 	if (!inlineConfig || !inlineConfig.projectRoot) throw new Error("Must provide { projectRoot: './fixtures/...' }");
@@ -63,10 +73,10 @@ export async function loadFixture(inlineConfig) {
 	return {
 		build: (opts = {}) => build(config, { mode: 'development', logging: 'error', ...opts }),
 		startDevServer: async (opts = {}) => {
-			const devServer = await dev(config, { logging: 'error', ...opts });
-			config.devOptions.port = devServer.port; // update port
-			inlineConfig.devOptions.port = devServer.port;
-			return devServer;
+			const devResult = await dev(config, { logging: 'error', ...opts });
+			config.devOptions.port = devResult.address.port; // update port
+			inlineConfig.devOptions.port = devResult.address.port;
+			return devResult;
 		},
 		config,
 		fetch: (url, init) => fetch(`http://${config.devOptions.hostname}:${config.devOptions.port}${url.replace(/^\/?/, '/')}`, init),
@@ -77,6 +87,7 @@ export async function loadFixture(inlineConfig) {
 		},
 		readFile: (filePath) => fs.promises.readFile(new URL(filePath.replace(/^\//, ''), config.dist), 'utf8'),
 		readdir: (fp) => fs.promises.readdir(new URL(fp.replace(/^\//, ''), config.dist)),
+		clean: () => fs.promises.rm(config.dist, { maxRetries: 10, recursive: true, force: true }),
 	};
 }